move ar experience to different component
